atlas: refactored zoneimport; introduced utils package
